isolate deltalake dependency
