+++ crates/lgn-analytics-srv/Cargo.toml
@@ -4,7 +4,9 @@ version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+deltalake-proto = ["dep:deltalake"]
+
 
 [dependencies]
 lgn-analytics = { path = "../lgn-analytics", version = "0.1.0" }
@@ -22,7 +24,7 @@ bytes = "1"
 chrono = "0.4"
 clap = { version = "3.0", features = ["derive"] }
 blake3 = { version = "1.3", features = ["prefer_intrinsics"] }
-deltalake = { git = "https://github.com/delta-io/delta-rs.git", rev = "9216a8e2f068cba7c7a965f51ab608ee1f55d773", version = "0.4.1", features = ["datafusion-ext"] }
+deltalake = { git = "https://github.com/delta-io/delta-rs.git", rev = "9216a8e2f068cba7c7a965f51ab608ee1f55d773", version = "0.4.1", features = ["datafusion-ext"], optional = true }
 datafusion = { git = "https://github.com/apache/arrow-datafusion.git", rev = "f943b6a65291c3de6403dde827cb411a050f4fa0", version = "7.0.0" }
 futures = "0.3"
 http = "0.2"
