+++ crates/lgn-analytics-srv/src/lakehouse/jit_lakehouse.rs
@@ -5,6 +5,7 @@ use lgn_telemetry_proto::analytics::BlockSpansReply;
 #[async_trait]
 pub trait JitLakehouse: Send + Sync {
     // build_timeline_tables is for prototyping the use of deltalake
+    #[cfg(feature = "deltalake-proto")]
     async fn build_timeline_tables(&self, process_id: &str) -> Result<()>;
 
     async fn get_thread_block(
