+++ crates/lgn-analytics-srv/src/lakehouse/mod.rs
@@ -1,3 +1,4 @@
+pub mod bytes_chunk_reader;
 pub mod column;
 pub mod jit_lakehouse;
 pub mod local_jit_lakehouse;
@@ -5,4 +6,9 @@ pub mod parquet_buffer;
 pub mod remote_jit_lakehouse;
 pub mod scope_table;
 pub mod span_table;
+
+#[cfg(feature = "deltalake-proto")]
 pub mod span_table_partition;
+
+#[cfg(feature = "deltalake-proto")]
+pub mod span_delta_table;
